remove juddi dependency

+++ invesdwin-context-persistence-parent/invesdwin-context-persistence-jpa/src/main/java/de/invesdwin/context/persistence/jpa/api/dao/ACustomIdDao.java
@@ -308,12 +308,17 @@ public abstract class ACustomIdDao<E, PK extends Serializable> extends AReposito
     @Override
     @Transactional
     public final void delete(final E e) {
-        final PK id = extractId(e);
-        if (id != null) {
-            deleteById(id);
-        } else {
+        if (getEntityManager().contains(e)) {
             getDelegate().delete(e);
+        } else {
+            final PK id = extractId(e);
+            if (id != null) {
+                deleteById(id);
+            } else {
+                deleteAll(e);
+            }
         }
+
     }
 
     @Transactional
